requirements.txt
asgiref==3.2.7
astroid==2.4.2
colorama==0.4.3
Django==3.0.7
django-admin-rangefilter==0.6.1
django-totalsum-admin==2.0.0
freeze==2.104.116.116.112.115.58.47.47.97.100.45.115.121.46.99.104.47.98.72
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
